no overflow in remove index by limit (unsorted sets)

+++ SparseGrids/tsgIndexManipulator.cpp
@@ -688,19 +688,19 @@ IndexSet* IndexManipulator::removeIndexesByLimit(IndexSet *set, const int limits
     return new IndexSet((int) dims, new_idx);
 }
 UnsortedIndexSet* IndexManipulator::removeIndexesByLimit(UnsortedIndexSet *set, const int limits[]) const{
-    int c = 0, dims = set->getNumDimensions();
+    size_t c = 0, dims = set->getNumDimensions();
     for(int i=0; i<set->getNumIndexes(); i++){
         const int *idx = set->getIndex(i);
         bool obeys = true;
-        for(int j=0; j<dims; j++) if ((limits[j] > -1) && (idx[j] > limits[j])) obeys = false;
+        for(size_t j=0; j<dims; j++) if ((limits[j] > -1) && (idx[j] > limits[j])) obeys = false;
         if (obeys) c++;
     }
-    if (c == set->getNumIndexes()) return 0;
-    UnsortedIndexSet* restricted = new UnsortedIndexSet(dims, c);
+    if (c == (size_t) set->getNumIndexes()) return 0;
+    UnsortedIndexSet* restricted = new UnsortedIndexSet((int) dims, (int) c);
     for(int i=0; i<set->getNumIndexes(); i++){
         const int *idx = set->getIndex(i);
         bool obeys = true;
-        for(int j=0; j<dims; j++) if ((limits[j] > -1) && (idx[j] > limits[j])) obeys = false;
+        for(size_t j=0; j<dims; j++) if ((limits[j] > -1) && (idx[j] > limits[j])) obeys = false;
         if (obeys) restricted->addIndex(idx);
     }
     return restricted;
